requirements.txt
pandas==2.2.2
numpy==1.26.4
prophet==1.1.5
streamlit==1.34.0
plotly==5.22.0
python-dotenv==1.0.1
phi==0.4.2
groq==0.3.0
pytest==8.2.2
